Remove not useful params to create block
